Feature: Support multiple beta build localizations for TestFlight publishing (#124)
**Summary of the changes:**

The action `app-store-connect publish` already allowed submitting release notes, or _what's new_ notes in App Store Connect UI terms, to uploaded builds. But there was a limitation that you could only specify the notes for a single language using `--whats-new` flag. Often times this is not enough and you'd want to automate the full process of beta release cycle for whatever languages/locales you have. With the changes introduced in this pull request `app-store-connect publish` gets a new option `--beta-build-localizations` that takes JSON encoded list of locales and notes (as seen below in the example) and automates release notes creation for all given languages.

**Related changes:**

- Make `--whats-new` option independent from `--testflight` for `app-store-connect publish` as submission to external beta review is not required for specifying notes.
- Do not print out found `PreReleaseVersion` of uploaded build as it carries little extra information to the users.
- Allow plain `str` input for `whats_new` arguments to actions defined in `beta_build_localizations_action_group.py`. 
- Add more blue and green colors to logs to indicate the start of an activity and completion of it.
- Use `primaryLocale` attribute from application to detect the default locale instead of using the first `betaAppLocalization` for that app.
- Show more descriptive error messages for invalid inputs to CLI arguments that can be defined using `@env:<...>` and `@file:<...>` notations.